Run benchmarks on updated datasets
Two updates:
- Use GTDB release 07-r207 genomes
- Switch from L. plantarum to the genus Lactiplantibacillus

+++ src/01_prepare_lactobacillales/01_select_representatives.R
@@ -1,40 +1,43 @@
 #!/usr/bin/env Rscript
 
-# This script will select GenBank accession numbers of species and genus 
-# representatives of the order Lactobacillales. The best-quality genome per 
-# species/genus is chosen as representative. As a consequence, the genus 
-# representatives are a subset of the species representatives. 
+# This script selects GenBank accession numbers of species and genus
+# representatives of the order Lactobacillales. The best-quality genome per
+# species/genus is chosen as representative. As a consequence, the genus
+# representatives are a subset of the species representatives.
 
-# dependencies: R v3.6.1, tidyverse v1.2.1
+# depencencies: R v4.1.2, tidyverse v2.0.0
 
 library(tidyverse)
 
 # define paths
-url_gtdb <- "https://data.ace.uq.edu.au/public/gtdb/data/releases/release89/89.0/bac120_metadata_r89.tsv"
-fout_metadata <- "../../data/lactobacillales/bac120_metadata_r89.tsv"
+url_gtdb <- "https://data.gtdb.ecogenomic.org/releases/release207/207.0/bac120_metadata_r207.tar.gz" # GTDB07
 dout <- "../../data/lactobacillales"
 
 # define derived paths
-fout_metadata <- paste0(dout, "/bac120_metadata_r89.tsv")
+fout_metadata_raw <- paste0(dout, "/bac120_metadata_r207.tar.gz")
+fout_metadata <- paste0(dout, "/bac120_metadata_r207.tsv")
 fout_accessions_species <- paste0(dout, "/accessions_speciesreps.tsv")
 fout_accessions_genera <- paste0(dout, "/accessions_genusreps.tsv")
 
 # create output folder
 if (! dir.exists(dout)) dir.create(dout)
 
-# download genome metadata file from the GTDB
+# download genome metadata file from the GTDB and untar
+if (! file.exists(fout_metadata_raw)) {
+  download.file(url_gtdb, destfile = fout_metadata_raw)
+}
 if (! file.exists(fout_metadata)) {
-  download.file(url_gtdb, destfile = fout_metadata)
+  untar(fout_metadata_raw, exdir = dout)
 }
 
 # read GTDB genomes and select Lactobacillales
-genomes <- 
-  read_tsv(fout_metadata) %>%
-  mutate(
-    gtdb_order = str_extract(gtdb_taxonomy, "o__[a-zA-Z_]+"),
-    gtdb_genus = str_extract(gtdb_taxonomy, "g__[a-zA-Z_]+"),
-    gtdb_species = str_extract(gtdb_taxonomy, "s__[a-zA-Z_ ]+")
-  ) %>%
+ranks <- c(
+  "gtdb_domain", "gtdb_phylum", "gtdb_class", "gtdb_order", "gtdb_family", 
+  "gtdb_genus", "gtdb_species"
+)
+genomes <-
+  read_tsv(fout_metadata, col_types = cols()) %>%
+  separate(gtdb_taxonomy, into = ranks, sep = ";") %>%
   filter(gtdb_order == "o__Lactobacillales") %>%
   filter(ncbi_genbank_assembly_accession != "none") %>%
   mutate(checkm_quality = checkm_completeness - checkm_contamination)
@@ -58,4 +61,7 @@ genomes %>%
   write_tsv(fout_accessions_genera, col_names = F)
 
 # remove large file with raw genome metadata
-file.remove(fout_metadata)
+if (file.exists(fout_accession_genera)) {
+  file.remove(fout_metadata_raw)
+  file.remove(fout_metadata)
+}
